+++ jest.config.ts
@@ -188,9 +188,10 @@ const config: Config = {
     // ],
 
     // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
-    // testPathIgnorePatterns: [
-    //   "\\\\node_modules\\\\"
-    // ],
+    testPathIgnorePatterns: ['\\\\node_modules\\\\'],
+    moduleNameMapper: {
+        '\\.(css|less|scss)$': 'identity-obj-proxy',
+    },
 
     // The regexp pattern or array of patterns that Jest uses to detect test files
     // testRegex: [],
