+++ src/components/Button/Button.tsx
@@ -1,27 +1,8 @@
-import { VariantProps, tv } from 'tailwind-variants';
 import { ComponentProps, forwardRef } from 'react';
 import '@govbr-ds/core/dist/components/button/button.min.css';
+import { buttonVariants, ButtonVariants } from './variants';
 
 type Ref = HTMLButtonElement;
-
-const buttonVariants = tv({
-    base: 'br-button',
-    variants: {
-        variant: {
-            primary: 'primary',
-            outline: 'secondary',
-            transparent: '',
-        },
-        loading: {
-            true: 'loading',
-        },
-    },
-    defaultVariants: {
-        variant: 'primary',
-    },
-});
-
-type ButtonVariants = VariantProps<typeof buttonVariants>;
 
 export type ButtonProps = ButtonVariants & ComponentProps<'button'>;
