+++ src/components/Button/__tests__/Button.spec.tsx
@@ -1,8 +1,8 @@
 import { render } from '@testing-library/react';
-import { Button } from '.';
+import { Button } from '..';
 describe('Button', () => {
     it('should render', () => {
-        const { getByText } = render(<Button></Button>);
+        const { getByText } = render(<Button>Botão</Button>);
         expect(getByText('Botão')).toBeInTheDocument();
     });
 });
