+++ src/styles/globals.css
@@ -1,4 +1,5 @@
 @import 'variables.css';
+@import 'states.css';
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
